refactoring api: create camera endpoint
