+++ src/App.tsx
@@ -5,6 +5,7 @@ import Header from './components/Header/Header';
 import AdminPanelPage from './pages/AdminPanelPage';
 import CameraDetailsPage from './pages/CameraDetailsPage';
 import CamerasPage from './pages/CamerasPage';
+import FilmsPage from './pages/FilmsPage';
 import HomePage from './pages/HomePage';
 import RoutesPath from './routes';
 
@@ -17,6 +18,7 @@ const App = () => {
           <Route path={RoutesPath.HOME} element={<HomePage />} />
           <Route path={RoutesPath.CAMERAS} element={<CamerasPage />} />
           <Route path={RoutesPath.CAMERAID} element={<CameraDetailsPage />} />
+          <Route path={RoutesPath.FILMS} element={<FilmsPage />} />
 
           <Route path={RoutesPath.ADMINPANEL} element={<AdminPanelPage />} />
           <Route path="*" element={<HomePage />} />
