+++ src/pages/FilmsPage.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const FilmsPage: FC = () => {
+  return <>films page</>;
+};
+
+export default FilmsPage;
