+++ src/routes.ts
@@ -2,6 +2,7 @@ enum RoutesPath {
   HOME = '/',
   CAMERAS = 'camera',
   CAMERAID = 'camera/:id',
+  FILMS = 'film',
   ADMINPANEL = 'admin',
 }
